Move code away from Analyzer
