+++ detekt-api/src/main/kotlin/io/gitlab/arturbosch/detekt/api/SetupContext.kt
@@ -17,11 +17,6 @@ interface SetupContext : PropertiesAware {
     val config: Config
 
     /**
-     * Configuration which is used to set up detekt without default config behind
-     */
-    val baseConfig: Config
-
-    /**
      * The channel to log all the output.
      */
     val outputChannel: Appendable
