+++ detekt-api/src/testFixtures/kotlin/io/gitlab/arturbosch/detekt/test/EmptySetupContext.kt
@@ -7,7 +7,6 @@ import java.net.URI
 class EmptySetupContext : SetupContext {
     override val configUris: Collection<URI> = emptyList()
     override val config: Config = Config.empty
-    override val baseConfig: Config = Config.empty
     override val outputChannel: Appendable = StringBuilder()
     override val errorChannel: Appendable = StringBuilder()
     override val properties: MutableMap<String, Any?> = HashMap()
