+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/ProcessingSettings.kt
@@ -4,7 +4,6 @@ import io.github.detekt.tooling.api.spec.ProcessingSpec
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.PropertiesAware
 import io.gitlab.arturbosch.detekt.api.SetupContext
-import io.gitlab.arturbosch.detekt.core.config.CompositeConfig
 import io.gitlab.arturbosch.detekt.core.config.extractUris
 import io.gitlab.arturbosch.detekt.core.settings.ClassloaderAware
 import io.gitlab.arturbosch.detekt.core.settings.EnvironmentAware
@@ -13,7 +12,6 @@ import io.gitlab.arturbosch.detekt.core.settings.ExtensionFacade
 import io.gitlab.arturbosch.detekt.core.settings.LoggingAware
 import io.gitlab.arturbosch.detekt.core.settings.LoggingFacade
 import io.gitlab.arturbosch.detekt.core.settings.PropertiesFacade
-import io.gitlab.arturbosch.detekt.core.tooling.getDefaultConfiguration
 import org.jetbrains.kotlin.com.intellij.openapi.util.Disposer
 import org.jetbrains.kotlin.utils.closeQuietly
 import java.io.Closeable
@@ -27,7 +25,7 @@ import java.net.URI
  */
 class ProcessingSettings(
     val spec: ProcessingSpec,
-    override val baseConfig: Config,
+    override val config: Config,
 ) : AutoCloseable,
     Closeable,
     LoggingAware by LoggingFacade(spec.loggingSpec),
@@ -37,12 +35,6 @@ class ProcessingSettings(
     SetupContext {
 
     override val configUris: Collection<URI> = spec.configSpec.extractUris()
-    override val config: Config
-        get() = if (spec.configSpec.useDefaultConfig) {
-            CompositeConfig(baseConfig, spec.getDefaultConfiguration())
-        } else {
-            baseConfig
-        }
 
     /**
      * Sharable thread pool between parsing and analysis phase.
